问题优化，最新APK生成

+++ app/src/main/java/com/example/medicalservice/bean/CanteenBean.java
@@ -58,7 +58,7 @@ public class CanteenBean {
         private String rFoodCanteenRemark;
         private String userId;
         private String userType;
-        private int markerId;
+        private Long markerId;
         private String markerName;
 
         public String getCreateBy() {
@@ -181,11 +181,11 @@ public class CanteenBean {
             this.userType = userType;
         }
 
-        public int getMarkerId() {
+        public Long getMarkerId() {
             return markerId;
         }
 
-        public void setMarkerId(int markerId) {
+        public void setMarkerId(Long markerId) {
             this.markerId = markerId;
         }
